+++ scrypto/src/component/system.rs
@@ -1,56 +1,11 @@
 use crate::buffer::*;
 use crate::component::package::Package;
 use crate::component::*;
-use crate::core::{DataAddress, DataRef, SNodeRef};
+use crate::core::SNodeRef;
 use crate::engine::{api::*, call_engine};
-use crate::prelude::DataRefMut;
 use sbor::rust::borrow::ToOwned;
-use sbor::rust::cell::{Ref, RefCell, RefMut};
 use sbor::rust::collections::*;
 use sbor::rust::string::ToString;
-use sbor::{Decode, Encode};
-
-pub struct DataValue<V: 'static + Encode + Decode> {
-    address: DataAddress,
-    value: RefCell<Option<V>>,
-}
-
-impl<V: 'static + Encode + Decode> DataValue<V> {
-    pub fn new(address: DataAddress) -> Self {
-        Self {
-            address,
-            value: RefCell::new(Option::None),
-        }
-    }
-
-    /// Returns a reference to component data
-    pub fn get_mut(&mut self) -> DataRefMut<V> {
-        let value = RefMut::map(self.value.borrow_mut(), |v| {
-            v.get_or_insert_with(|| {
-                let input =
-                    ::scrypto::engine::api::RadixEngineInput::ReadData(self.address.clone());
-                let value: V = call_engine(input);
-                value
-            })
-        });
-        DataRefMut {
-            address: self.address.clone(),
-            value,
-        }
-    }
-
-    pub fn get(&self) -> DataRef<V> {
-        self.value.borrow_mut().get_or_insert_with(|| {
-            let input = ::scrypto::engine::api::RadixEngineInput::ReadData(self.address.clone());
-            let value: V = call_engine(input);
-            value
-        });
-
-        let value = Ref::map(self.value.borrow(), |v| v.as_ref().unwrap());
-
-        DataRef { value }
-    }
-}
 
 /// Represents the Radix Engine component subsystem.
 ///
